+++ WebAPIMatricula_3C2023/API.Dto.Fin/Entrada/AgregarFinanza.cs
@@ -9,9 +9,9 @@ namespace API.Dto.Finanza.Entrada
     public class AgregarFinanza : General.EntradaAPI
     {
         public int TotalMaterias { get; set; }
-        public int CobroMatricula { get; set; }
-        public int CobroPoliza { get; set; }
-        public int CobroTechFee { get; set; }
+        public float CobroMatricula { get; set; }
+        public float CobroPoliza { get; set; }
+        public float CobroTechFee { get; set; }
         public int CodigoMatricula { get; set; }
     }
 }
